+++ weconnect/util.py
@@ -4,6 +4,8 @@ from typing import Any
 import re
 from datetime import datetime
 
+import logging
+
 import shutil
 
 from PIL import Image  # type: ignore
@@ -54,3 +56,27 @@ def celsiusToKelvin(value):
 
 def farenheitToKelvin(value):
     return 273.5 + ((value - 32.0) * (5.0 / 9.0))
+
+
+class DuplicateFilter(logging.Filter):
+    def __init__(self, name: str = ...) -> None:
+        super().__init__(name=name)
+        self.lastLog = {}
+        self.firstTime = True
+
+    def filter(self, record) -> bool:
+        # don't filter critical or error messages:
+        if record.levelno >= logging.ERROR:
+            return True
+
+        if record.module in self.lastLog:
+            if record.levelno in self.lastLog[record.module]:
+                if self.lastLog[record.module][record.levelno] == (record.msg, record.args):
+                    if self.firstTime:
+                        self.firstTime = False
+                        logging.info('Repeated log messages from the same module are hidden (does not apply to errors or critical problems)')
+                    return False
+            self.lastLog[record.module][record.levelno] = (record.msg, record.args)
+        else:
+            self.lastLog[record.module] = {record.levelno: (record.msg, record.args)}
+        return True