+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 ### Added
 - Added MBB Platform (used in e.g. e-up)
+- Duplicate log entry filter
 
 ### Changed
 - Changed to new selective update endpoint. Warning this is a breaking change in the API!